pyproject.toml
[project]
name = "tradenote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "joblib==1.1.0",
    "jupyter>=1.1.1",
    "notebook>=7.4.3",
    "plotly>=6.1.2",
    "pyqlib>=0.9.6",
]
